Fixed race condition for async workers with ai api job and improved report design

+++ src/main/java/com/ua/pohribnyi/fitadvisorbot/model/dto/google/DailyMetricDto.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 import com.ua.pohribnyi.fitadvisorbot.model.entity.DailyMetric;
 import com.ua.pohribnyi.fitadvisorbot.model.entity.user.User;
 
-public record DailyMetricDto(String date, Float sleepHours, Integer dailyBaseSteps, Integer stressLevel) {
+public record DailyMetricDto(String date, Double sleepHours, Integer dailyBaseSteps, Integer stressLevel) {
 
 	public static DailyMetric mapToEntity(DailyMetricDto dto, User user) {
 		DailyMetric entity = new DailyMetric();
